RANGER-3905: Ranger metrics on audit throughput for each of the Ranger Service plugins - fix review comment
