+++ agents-common/src/main/java/org/apache/ranger/plugin/util/SearchFilter.java
@@ -96,7 +96,7 @@ public class SearchFilter {
     public static final String SERVICE_NAME_PARTIAL = "serviceNamePartial";
 
     public static final String CLIENT_IP        = "clientIP";
-    public static final String APP_ID           = "appId";
+    public static final String PLUGIN_APP_ID    = "pluginAppId";
 
     public static final String PLUGIN_HOST_NAME           = "pluginHostName";
     public static final String PLUGIN_APP_TYPE            = "pluginAppType";
